refactor: move renderers to shared directory
